refactor: create plugin config
