--- src/handlers/main.ts
@@ -1,11 +1,0 @@
-import { Distributor } from './distributor';
-import { engines } from '@txtdot/plugins';
-
-const distributor = new Distributor();
-
-distributor.engine(engines.StackOverflow);
-distributor.engine(engines.SearX);
-distributor.engine(engines.Readability);
-
-export const engineList = distributor.list;
-export default distributor;
