+++ src/types/appConfig.ts
@@ -1,0 +1,5 @@
+import { Engine } from '@txtdot/sdk';
+
+export interface IAppConfig {
+  engines: Engine[];
+}
