+++ src/types/requests/api.ts
@@ -1,6 +1,6 @@
 import { FastifySchema, FastifyRequest } from 'fastify';
 import { IApiError, errorResponseSchema } from '../../errors/api';
-import { engineList } from '../../handlers/main';
+import { engineList } from '../../plugin_manager';
 import { FromSchema } from 'json-schema-to-ts';
 import { handlerSchema } from '@txtdot/sdk/dist/types/handler';
